requirements.txt
python-telegram-bot>=20.0
asyncio
asyncpg>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
structlog>=23.0.0